Added a script to make zellij start temporary sessions when launched through SUPER+Return. Added bind for this in hyprland.
The script deletes all non-running scripts that contain temp in their
name, which ensures that all trash sessions are removed quickly.
The point is that I would like to have zellij active in all terminals,
but not every session is worth keeping. This way I can rename the
sessions I would like to restore, and the rest automatically get cleaned
up.