Waybar: Updated and simplified config
It now looks a lot cleaner and better, and uses colors from everforest.
I removed diagnostics, as I never really need to know my CPU and memory usage.
If I need these things, mangohud does a better job.

Background color is a slightly brown grey, set to be slightly transparent.
All borders are turned off, and font size is small for a cleaner look.
The goal here is to later add more information (schedule most likely),
but really to just make a bar that is so minimal that I will keep it shown
and not just hide it on every boot.